Integrate Redux dev tools.
