+++ package.json
@@ -111,6 +111,10 @@
     "react-transform-hmr": "^1.0.2",
     "react-transform-render-visualizer": "^0.4.0",
     "redbox-react": "^1.2.2",
+    "redux-devtools": "^3.1.1",
+    "redux-devtools-dock-monitor": "^1.1.0",
+    "redux-devtools-log-monitor": "^1.0.5",
+    "redux-slider-monitor": "^1.0.2",
     "reporter": "^0.1.0",
     "rimraf": "^2.4.3",
     "sinon": "^1.17.3",
